refactor: split json explorer sidebar sections
