+++ src/pages/json-explorer/features/JsonExplorer/ui/JsonExplorerSidebar.tsx
@@ -1,13 +1,9 @@
-/**
- * JsonExplorerSidebar - JSON 키 경로 탐색 사이드바
- * DataSource (데이터 소스 선택) + Columns (컬럼 목록) + Schema (키 경로 트리) 세 섹션으로 구성
- */
-
-import { ChevronDown, ChevronRight, Columns, Copy, Database, FileJson, GitBranch, Pencil } from 'lucide-react';
-import { useMemo, useState } from 'react';
-import { ScrollArea } from '@/shared/ui/ScrollArea';
-import { extractSchemaInterfaces, type SchemaInterfaceNode } from '../lib/extractKeyPaths';
+import { useMemo } from 'react';
+import { extractSchemaInterfaces } from '../lib/extractKeyPaths';
 import type { DataSource } from '../lib/loadTestData';
+import { JsonExplorerColumnsSection } from './JsonExplorerColumnsSection';
+import { JsonExplorerDataSourceSection } from './JsonExplorerDataSourceSection';
+import { JsonExplorerSchemaSection } from './JsonExplorerSchemaSection';
 
 interface JsonExplorerSidebarProps {
   columns: string[]; // 1depth 컬럼 목록
@@ -19,202 +15,6 @@ interface JsonExplorerSidebarProps {
   onCustomJsonClick: () => void; // 커스텀 JSON 입력 모달 열기
   selectedSchema: string | null; // 선택된 스키마 인터페이스 이름
   onSchemaSelect: (schemaPath: string | null) => void; // 스키마 선택 콜백
-}
-
-/**
- * Schema Interface Item - TypeScript interface 형식 렌더링
- */
-function SchemaInterfaceItem({
-  node,
-  isSelected,
-  onSelect,
-}: {
-  node: SchemaInterfaceNode;
-  isSelected: boolean;
-  onSelect: (path: string | null) => void;
-}) {
-  const [isExpanded, setIsExpanded] = useState(false);
-  const [copiedPath, setCopiedPath] = useState<string | null>(null);
-
-  async function handleCopyPath(path: string, e: React.MouseEvent) {
-    e.stopPropagation();
-    try {
-      await navigator.clipboard.writeText(path);
-      setCopiedPath(path);
-      setTimeout(() => setCopiedPath(null), 2000);
-    } catch (error) {
-      console.error('Failed to copy path:', error);
-    }
-  }
-
-  async function handleCopyInterface(e: React.MouseEvent) {
-    e.stopPropagation();
-    try {
-      // TypeScript interface 코드 생성
-      const interfaceCode = `interface ${node.interfaceName} {\n${node.fields
-        .map((field) => {
-          const displayType = field.isArray ? `${field.type}[]` : field.type;
-          const optional = field.isOptional ? '?' : '';
-          return `  ${field.name}${optional}: ${displayType};`;
-        })
-        .join('\n')}\n}`;
-      await navigator.clipboard.writeText(interfaceCode);
-      setCopiedPath('interface');
-      setTimeout(() => setCopiedPath(null), 2000);
-    } catch (error) {
-      console.error('Failed to copy interface:', error);
-    }
-  }
-
-  function handleToggle() {
-    if (isSelected) {
-      // 이미 선택된 상태면 선택 해제
-      onSelect(null);
-      setIsExpanded(false);
-    } else {
-      // 선택
-      onSelect(node.path);
-      setIsExpanded(true);
-    }
-  }
-
-  function handleFieldCopyPath(path: string, e: React.MouseEvent) {
-    handleCopyPath(path, e);
-  }
-
-  return (
-    <div className="mb-2">
-      {/* Interface Header */}
-      <div
-        className={`flex items-center gap-1.5 px-2 py-1 cursor-pointer hover:bg-warm-500/10 transition-colors group ${
-          isSelected ? 'bg-warm-500/20' : ''
-        }`}
-        onClick={handleToggle}
-      >
-        {isExpanded ? (
-          <ChevronDown size={10} className="shrink-0 text-text-tertiary" />
-        ) : (
-          <ChevronRight size={10} className="shrink-0 text-text-tertiary" />
-        )}
-        <span className={`text-2xs font-semibold font-mono ${isSelected ? 'text-warm-400' : 'text-warm-300'}`}>
-          interface {node.interfaceName}
-        </span>
-        <button
-          onClick={handleCopyInterface}
-          className="ml-auto opacity-0 group-hover:opacity-100 transition-opacity p-0.5 hover:bg-bg-elevated rounded"
-          title="Copy interface"
-          aria-label="Copy interface"
-        >
-          <Copy size={10} className={copiedPath === 'interface' ? 'text-green-400' : 'text-text-tertiary'} />
-        </button>
-      </div>
-
-      {/* Path (collapsed) */}
-      {!isExpanded && (
-        <div className="text-2xs text-text-tertiary font-mono italic px-2 pl-7 pb-1">{node.path || 'root'}</div>
-      )}
-
-      {/* Interface Body (expanded) */}
-      {isExpanded && (
-        <div className="pl-7 pr-2">
-          {/* Opening brace */}
-          <div className="text-2xs text-text-tertiary font-mono">{'{'}</div>
-
-          {/* Fields */}
-          {node.fields.map((field) => {
-            const displayType = field.isArray ? `${field.type}[]` : field.type;
-            const optional = field.isOptional ? '?' : '';
-
-            // 타입이 primitive인지 확인 (string, number, boolean, null)
-            const isPrimitive = ['string', 'number', 'boolean', 'null', 'unknown'].some(
-              (primitiveType) => field.type === primitiveType || field.type.includes(primitiveType)
-            );
-
-            // 중첩 객체/배열 타입은 다른 색상 (cyan-400), primitive는 기본 색상 (text-secondary)
-            const typeColor = isPrimitive ? 'text-text-secondary' : 'text-cyan-400';
-
-            return (
-              <SchemaFieldRow
-                key={field.name}
-                fieldName={field.name}
-                fullPath={field.fullPath}
-                optional={optional}
-                displayType={displayType}
-                typeColor={typeColor}
-                copiedPath={copiedPath}
-                onCopyPath={handleFieldCopyPath}
-              />
-            );
-          })}
-
-          {/* Closing brace */}
-          <div className="text-2xs text-text-tertiary font-mono">{'}'}</div>
-
-          {/* Full path */}
-          <div className="text-2xs text-warm-400/60 font-mono italic mt-1">
-            {'// '}
-            {node.path || 'root'}
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function SchemaFieldRow({
-  fieldName,
-  fullPath,
-  optional,
-  displayType,
-  typeColor,
-  copiedPath,
-  onCopyPath,
-}: {
-  fieldName: string;
-  fullPath: string;
-  optional: string;
-  displayType: string;
-  typeColor: string;
-  copiedPath: string | null;
-  onCopyPath: (path: string, e: React.MouseEvent) => void;
-}) {
-  function handleCopyPathClick(e: React.MouseEvent) {
-    onCopyPath(fullPath, e);
-  }
-
-  return (
-    <div className="flex items-center gap-1.5 py-0.5 hover:bg-warm-500/5 group">
-      <span className="text-2xs font-mono text-warm-400 pl-2">
-        {fieldName}
-        {optional}:
-      </span>
-      <span className={`text-2xs font-mono ${typeColor}`}>{displayType};</span>
-      <button
-        onClick={handleCopyPathClick}
-        className="ml-auto opacity-0 group-hover:opacity-100 transition-opacity p-0.5 hover:bg-bg-elevated rounded shrink-0"
-        title={`Copy path: ${fullPath}`}
-        aria-label="Copy path"
-      >
-        <Copy size={9} className={copiedPath === fullPath ? 'text-green-400' : 'text-text-tertiary'} />
-      </button>
-    </div>
-  );
-}
-
-function ColumnItem({ column, onSelectColumn }: { column: string; onSelectColumn: (column: string) => void }) {
-  function handleClick() {
-    onSelectColumn(column);
-  }
-
-  return (
-    <div
-      className="flex items-center gap-2 px-3 py-0.5 text-2xs cursor-pointer hover:bg-warm-500/10 transition-colors"
-      onClick={handleClick}
-    >
-      <div className="w-1 h-1 rounded-full bg-warm-300 shrink-0" />
-      <span className="font-mono text-text-secondary truncate">{column}</span>
-    </div>
-  );
 }
 
 export function JsonExplorerSidebar({
@@ -233,165 +33,20 @@ export function JsonExplorerSidebar({
     console.log('[JsonExplorerSidebar] Schema interfaces extracted:', interfaces.length, 'interfaces');
     return interfaces;
   }, [allData]);
-  const [dataSourceExpanded, setDataSourceExpanded] = useState(true);
-  const [schemaExpanded, setSchemaExpanded] = useState(true);
-  const [columnsExpanded, setColumnsExpanded] = useState(true);
-
-  function handleColumnClick(columnKey: string) {
-    onSelectPath?.(columnKey);
-    onScrollToColumn?.(columnKey);
-  }
-
-  function handleDataSourceToggle() {
-    setDataSourceExpanded(!dataSourceExpanded);
-  }
-
-  function handleTestJsonClick() {
-    onDataSourceChange('test.json');
-  }
-
-  function handleTest2JsonClick() {
-    onDataSourceChange('test2.json');
-  }
-
-  function handleCustomJsonClick() {
-    onCustomJsonClick();
-  }
-
-  function handleSchemaToggle() {
-    setSchemaExpanded(!schemaExpanded);
-  }
-
-  function handleSchemaSelect(schemaPath: string | null) {
-    onSchemaSelect(schemaPath);
-  }
-
-  function handleColumnsToggle() {
-    setColumnsExpanded(!columnsExpanded);
-  }
 
   return (
     <div className="w-64 border-r border-border-DEFAULT bg-bg-elevated flex flex-col h-full">
-      {/* DataSource Section */}
-      <div className="border-b border-border-DEFAULT">
-        {/* Section Header */}
-        <div
-          className="flex items-center gap-2 px-3 py-2 bg-bg-deep cursor-pointer hover:bg-bg-deep/80 transition-colors"
-          onClick={handleDataSourceToggle}
-        >
-          {dataSourceExpanded ? <ChevronDown size={12} /> : <ChevronRight size={12} />}
-          <Database size={12} className="text-warm-300" />
-          <h2 className="text-xs font-semibold text-text-primary">Data Source</h2>
-        </div>
-
-        {/* DataSource Options */}
-        {dataSourceExpanded && (
-          <div className="py-1">
-            {/* test.json */}
-            <div
-              className={`flex items-center gap-2 px-3 py-1.5 text-2xs cursor-pointer transition-colors ${
-                dataSource === 'test.json'
-                  ? 'bg-warm-500/20 text-text-primary'
-                  : 'text-text-secondary hover:bg-warm-500/10'
-              }`}
-              onClick={handleTestJsonClick}
-            >
-              <div
-                className={`w-1 h-1 rounded-full shrink-0 ${dataSource === 'test.json' ? 'bg-warm-400' : 'bg-text-tertiary'}`}
-              />
-              <FileJson size={12} className="shrink-0" />
-              <span className="font-mono truncate">test.json</span>
-            </div>
-
-            {/* test2.json */}
-            <div
-              className={`flex items-center gap-2 px-3 py-1.5 text-2xs cursor-pointer transition-colors ${
-                dataSource === 'test2.json'
-                  ? 'bg-warm-500/20 text-text-primary'
-                  : 'text-text-secondary hover:bg-warm-500/10'
-              }`}
-              onClick={handleTest2JsonClick}
-            >
-              <div
-                className={`w-1 h-1 rounded-full shrink-0 ${dataSource === 'test2.json' ? 'bg-warm-400' : 'bg-text-tertiary'}`}
-              />
-              <FileJson size={12} className="shrink-0" />
-              <span className="font-mono truncate">test2.json</span>
-            </div>
-
-            {/* Custom JSON */}
-            <div
-              className={`flex items-center gap-2 px-3 py-1.5 text-2xs cursor-pointer transition-colors ${
-                dataSource === 'custom'
-                  ? 'bg-warm-500/20 text-text-primary'
-                  : 'text-text-secondary hover:bg-warm-500/10'
-              }`}
-              onClick={handleCustomJsonClick}
-            >
-              <div
-                className={`w-1 h-1 rounded-full shrink-0 ${dataSource === 'custom' ? 'bg-warm-400' : 'bg-text-tertiary'}`}
-              />
-              <Pencil size={12} className="shrink-0" />
-              <span className="truncate">Custom JSON...</span>
-            </div>
-          </div>
-        )}
-      </div>
-
-      {/* Schema Section */}
-      <div className="border-b border-border-DEFAULT flex flex-col overflow-hidden">
-        {/* Section Header */}
-        <div
-          className="flex items-center gap-2 px-3 py-2 bg-bg-deep cursor-pointer hover:bg-bg-deep/80 transition-colors shrink-0"
-          onClick={handleSchemaToggle}
-        >
-          {schemaExpanded ? <ChevronDown size={12} /> : <ChevronRight size={12} />}
-          <GitBranch size={12} className="text-warm-300" />
-          <h2 className="text-xs font-semibold text-text-primary">Schema</h2>
-          <span className="text-3xs text-text-tertiary ml-auto">{schemaInterfaces.length} interfaces</span>
-        </div>
-
-        {/* Schema Interfaces */}
-        {schemaExpanded && (
-          <ScrollArea className="flex-1 max-h-[50vh]">
-            <div className="py-2">
-              {schemaInterfaces.map((interfaceNode) => (
-                <SchemaInterfaceItem
-                  key={interfaceNode.path}
-                  node={interfaceNode}
-                  isSelected={selectedSchema === interfaceNode.path}
-                  onSelect={handleSchemaSelect}
-                />
-              ))}
-            </div>
-          </ScrollArea>
-        )}
-      </div>
-
-      {/* Columns Section */}
-      <div className="flex-1 flex flex-col overflow-hidden">
-        {/* Section Header */}
-        <div
-          className="flex items-center gap-2 px-3 py-2 bg-bg-deep border-b border-border-DEFAULT cursor-pointer hover:bg-bg-deep/80 transition-colors shrink-0"
-          onClick={handleColumnsToggle}
-        >
-          {columnsExpanded ? <ChevronDown size={12} /> : <ChevronRight size={12} />}
-          <Columns size={12} className="text-warm-300" />
-          <h2 className="text-xs font-semibold text-text-primary">Columns</h2>
-          <span className="text-3xs text-text-tertiary ml-auto">{columns.length}</span>
-        </div>
-
-        {/* Columns List */}
-        {columnsExpanded && (
-          <ScrollArea className="flex-1">
-            <div className="py-1">
-              {columns.map((col) => (
-                <ColumnItem key={col} column={col} onSelectColumn={handleColumnClick} />
-              ))}
-            </div>
-          </ScrollArea>
-        )}
-      </div>
+      <JsonExplorerDataSourceSection
+        dataSource={dataSource}
+        changeDataSource={onDataSourceChange}
+        openCustomJson={onCustomJsonClick}
+      />
+      <JsonExplorerSchemaSection
+        schemaInterfaces={schemaInterfaces}
+        selectedSchema={selectedSchema}
+        selectSchema={onSchemaSelect}
+      />
+      <JsonExplorerColumnsSection columns={columns} selectColumn={onSelectPath} scrollToColumn={onScrollToColumn} />
     </div>
   );
 }
